5-sqrt_recursion.c

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -9,7 +9,7 @@
  */
 int _sqrt_recursion(int n)
 {
-int root;
+long int root;
 if (n == 0)
 {
 return (0);
